Update installer-$arch/current in stage 2
There are scripts fetching the current installer images through
the dists/$release/main/installer-$arch/current/images/ path
and they are validating the MD5SUMS and SHA256SUMS in this
directory against the Release file. So let's update the symlink
in stage 2 as well so that we ensure that everything is in sync
at this point.